# Conflicts:
#	aiurt-base/aiurt-base-core/src/main/java/com/aiurt/boot/constant/SysParamCodeConstant.java
#	aiurt-module-fault/aiurt-module-fault-biz/src/main/java/com/aiurt/modules/common/api/DailyFaultApiImpl.java
#	aiurt-module-fault/aiurt-module-fault-biz/src/main/java/com/aiurt/modules/fault/service/impl/FaultServiceImpl.java
#	aiurt-module-patrol/aiurt-module-patrol-biz/src/main/java/com/aiurt/boot/api/PatrolApiServiceImpl.java
#	aiurt-module-patrol/aiurt-module-patrol-biz/src/main/java/com/aiurt/boot/task/mapper/xml/PatrolTaskMapper.xml
#	aiurt-module-system/src/main/java/com/aiurt/modules/subsystem/service/impl/CsSubsystemServiceImpl.java
#	aiurt-module-system/src/main/java/com/aiurt/modules/system/mapper/xml/CsUserSubsystemMapper.xml

+++ aiurt-module-patrol/aiurt-module-patrol-api/src/main/java/com/aiurt/boot/dto/UserTeamPatrolDTO.java
@@ -21,8 +21,9 @@ public class UserTeamPatrolDTO {
     private String orgId;
     @ApiModelProperty(value = "班组Id")
     private String orgCode;
+    /**巡视工时 2023-06-12通信6期，单位改成秒*/
     @ApiModelProperty(value = "巡视工时")
-    private BigDecimal workHours;
+    private Integer workHours;
     @ApiModelProperty(value = "计划任务数")
     private Integer planTaskNumber;
     @ApiModelProperty(value = "实际完成任务数")
